ExProsemirror.Block.HTML (ex_prosemirror v0.3.2)

HTML blocks only contain attrs that defines:

html: html content to embed

JS Interop

When inserting an HTML block, ex_prosemirror will send you a js event. With this event, you'll be able to open your own modal to gather the HTML input.

document.querySelectorAll('.ex-prosemirror').forEach((el) => {
  el.addEventListener('insertPlaceholder', phoenixHook)
});

Then you will need to send back the content to ex_prosemirror. This is achieved by dispatching an event to the ExEditorView instance.

function phoenixHook(event) {
  this.dispatchEvent(
    new CustomEvent(
      'replacePlaceholder',
      {detail: {id: event.detail.id, tr: event.detail.tr, data: {html: "<div style='color: red;'>Hello World</div>"}}}
    )
  );
}

Types

t()

@type t() :: %ExProsemirror.Block.HTML{
  attrs: %ExProsemirror.Block.HTML.Attrs{html: String.t(), id: term()}
}