ExProsemirror.Block.Image (ex_prosemirror v0.3.0)

Image only contains attrs that defines:

src: Source of the image

alt: Alternate text for an image, if the image cannot be displayed.

title: Title of the image

JS Interop

When inserting an image, ex_prosemirror will send you a js event. With this event, you'll be able to open your own modal for image upload for example.

document.querySelectorAll('.ex-prosemirror').forEach((el) => {
  el.addEventListener('insertPlaceholder', phoenixHook)
});

Then you will need to send back the content to ex_prosemirror. This is achieved by dispatching an event to the ExEditorView instance.

function phoenixHook(event) {
  this.dispatchEvent(
    new CustomEvent(
      'replacePlaceholder',
      {detail: {id: event.detail.id, tr: event.detail.tr, data: {url: "img_src"}}}
    )
  );
}

Types

t()

Specs

t() :: %ExProsemirror.Block.Image{
  attrs: %ExProsemirror.Block.Image.Attrs{
    alt: String.t(),
    id: term(),
    src: String.t(),
    title: String.t()
  }
}