ExProsemirror.Block.Text (ex_prosemirror v0.2.0)

Prosemirror inline element. It's contained by most of ExProsemirror blocks like ExProsemirror.Block.Paragraph or ExProsemirror.Block.Heading.

When creating an ExProsemirror block you can use this module thanks to

embedded_prosemirror_content([text: ExProsemirror.Block.Text], array: true)

You can also set array to false if you want to inline the block.

embedded_prosemirror_content is defined in ExProsemirror.SchemaHelper and is imported with using ExProsemirror.Schema.

Types

t()

Specs

t() :: %ExProsemirror.Block.Text{
  marks: [
    ExProsemirror.Mark.Strong.t()
    | ExProsemirror.Mark.Em.t()
    | ExProsemirror.Mark.Underline.t()
  ],
  text: String.t()
}

Functions

extract_simple_text(text)

Returns the text of an ExProsemirror.Block.Text.

E.g

ExProsemirror.Block.Text.extract_simple_text(%ExProsemirror.Block.Text{text: "Hello elixir's friends"})
"Hello elixir's friends"