ExProsemirror.EctoHelper (ex_prosemirror v0.1.0)

Helpers for ecto.

This module is automatically import if you use ExProsemirror.

Functions

cast_prosemirror_fields(struct_or_changeset)

Cast prosemirror data struct.

Examples

struct_or_changeset
|> cast(attrs, some_fields_to_cast)
|> cast_prosemirror_fields()

embedded_prosemirror_field(mapped_types, opts \\ [])
(macro)

Add the PolymorphicField in your ecto schema.

Examples

Single element

embedded_prosemirror_field([text: ExProsemirror.Text], array: false)
# same as
embedded_prosemirror_field([text: ExProsemirror.Text])

Multiple elements

embedded_prosemirror_field([text: ExProsemirror.Text], array: true)

Options

array: boolean

The array option will configure PolymorphicEmbed automatically to be a list of your data OR a single element.