ExProsemirror.HTML.Form (ex_prosemirror v0.1.2)

Form helpers to generate HTML fields required by ProseMirror.

Using this component, ProseMirror will automatically be link to hidden input fields. These hidden inputs will be used by Phoenix.HTML.Form to send data to the backend.

Live synchronization over LiveView is not currently supported.

Options

marks: mark tags to use defined in your ProseMirror config

blocks: block tags to use defined in your ProseMirror config

Functions

prosemirror_editor(form, field, opts \\ [])

Generates the <div> that will be used by javascript to mount the ProseMirror component.

Usage

<%= prosemirror_editor(@form, :title) %>

See ExProsemirror.HTML.Form for options

prosemirror_hidden_input(form, field, opts \\ [])

Generates a hidden input field to store data for ProseMirror.

Uses Phoenix.HTML.Form.html.hidden_input/3 under the hood.

Usage

<%= prosemirror_hidden_input(@form, :title) %>

prosemirror_input(form, field, opts \\ [])

Generates inputs for ProseMirror.

Usages

<%= prosemirror_input @form, :body, id: "my-article-input" %>

iex> prosemirror_input(form, :body)