ExProsemirror.Mark.Link (ex_prosemirror v0.3.2)

Transform a text to a link.

inputs opts:

marks: [:link]

JS part

We do not provide any JS modal for this mark.

You can easily create your own menu item like the following code.

export const generateLinkMenuItem = (schema) => {
  if (!schema.marks.link) {
    return [];
  }

  return [new MenuItem({
    title: 'Add or remove link',
    icon: icons.link,
    activate(state) { return markActive(state, schema.marks.link); },
    enable(state) { return !state.selection.empty; },
    run(state, dispatch, view) {
      // What happen when user click on the menu item button.
      // You can open a modal, activate / disable the mark ...
      // Read more on https://prosemirror.net/

      if (markActive(state, schema.marks.link)) {
        toggleMark(schema.marks.link)(state, dispatch);
        return true;
      }

      const attrs = {
        title: "Create a link",
        href: "https://your-link-here.com"
      }

      toggleMark(schema.marks.link, attrs)(view.state, view.dispatch)
      view.focus()
    }
  })];
}

Then in the your menu, you can use this function to display the link button.

Types

href()

@type href() :: String.t()

t()

@type t() :: %ExProsemirror.Mark.Link{
  attrs: %ExProsemirror.Mark.Link.Attrs{href: href(), title: title()}
}

title()

@type title() :: String.t()

Functions

changeset(struct_or_changeset, attrs \\ %{}, opts \\ [])