ExProsemirror.ModifierHelper (ex_prosemirror v0.2.0)

Ecto schema helper that defines multiple macros to use inside an Ecto.Schema.

In contrast of ExProsemirror.Schema, this module should not be imported to your business schema. It MUST be used to create custom block / marks.

It helps you build custom blocks / marks for ExProsemirror.

This module is automatically import if you use ExProsemirror.

Functions

cast_prosemirror_content(struct_or_changeset, opts \\ [])

Cast prosemirror data struct.

Examples

struct_or_changeset
|> cast(attrs, some_fields_to_cast)
|> cast_prosemirror_content()

# or with opts

struct_or_changeset
|> cast(attrs, some_fields_to_cast)
|> cast_prosemirror_content(with: [text: {ExProsemirror.Block.Text, :changeset, [opts]}])

Block ExProsemirror.Block.Text will be cast using the changeset/3 with params

%ExProsemirror.Block.Text{...}

attrs (map of data to cast)

opts: opts to use in the changeset (e.g. allowed marks to cast defined in the ExProsemirror type.)

If a block/mark's type is not specified in the with opts, the changeset/2 will be use.

Options

with: define the changeset to apply, to send allowed marks we recommend to set this element to passed opts.

cast_prosemirror_marks(struct_or_changeset, opts \\ [])

embedded_prosemirror_content(mapped_types, opts \\ [])
(macro)

Add the PolymorphicField mark in your ecto schema.

Examples

  embedded_prosemirror_content([text: ExProsemirror.Block.Text])

Options

array: boolean that defines if we could have multiple elements.

Use macro ExProsemirror.SchemaHelper.embedded_prosemirror_field/3.

embedded_prosemirror_field(field_name, mapped_types, opts \\ [])
(macro)

Specs

embedded_prosemirror_field(:content | :marks, [module()], [{:array, boolean()}]) ::
  term()

Add the PolymorphicField in your ecto schema.

Examples

Single element

embedded_prosemirror_field(:content, [text: ExProsemirror.Block.Text], array: false)
# same as
embedded_prosemirror_field(:content, [text: ExProsemirror.Block.Text])

Multiple elements

embedded_prosemirror_field(:content, [text: ExProsemirror.Block.Text], array: true)

Options

array: boolean

The array option will configure PolymorphicEmbed automatically to be a list of your data OR a single element.

embedded_prosemirror_marks(mapped_types)
(macro)

Add the PolymorphicField mark in your ecto schema.

The embedded_field created accepts an array of marks.

Examples

  embedded_prosemirror_mark([strong: ExProsemirror.Mark.Strong])

Use macro ExProsemirror.SchemaHelper.embedded_prosemirror_field/3.